requirements.txt
loguru
scrapy
scrapy-playwright
